chore: rename folder eth -> evm
