+++ engine/src/lib.rs
@@ -20,4 +20,4 @@ pub mod witness;
 // Blockchains
 pub mod btc;
 pub mod dot;
-pub mod eth;
+pub mod evm;
